P9.3: sdhci.toml flesh-out + lockjaw-regs::sdhci emission
Flip Sdhci's regspec from emit=false (skeletal placeholder since
Phase 2) to emit=true with the full BCM2711 emmc2 register surface
the driver actually touches. Generates user/lockjaw-regs/src/sdhci.rs
(1568 lines) — typed cells over the SDHCI MMIO layout with cross-
checked register offsets, typed flag/field/enum decoders, and
construction-safe accessors. The driver migration in P9.4-P9.11
consumes this surface; P9.3 is the codegen-prep that makes it
possible.

What lands
----------

- `user/regspecs/sdhci.toml` (+598 lines): full 27-register surface.
  Per-feature highlights:

  - `kind = "aliased"` on sysaddr (P9.1) — argument2 listed in the
    aliases array; codegen emits delegating accessors over one
    Rw<u32> cell field.

  - `kind = "combined_trigger"` on transfer_mode_command (P9.0):
    parts = [transfer_mode (15:0), command (31:16)]. The setter
    composes both parts into one Wo<u32>::write (M7 single-store
    fix structural). Typed per-part newtypes (P9.0b) wrap each
    arg so positional swap is a compile error.

  - `access = "rw1c"` on normal_int_status and error_int_status
    (this commit's substrate addition — see below). Codegen emits
    BOTH a typed snapshot read AND a typed clear_* mask write so
    the driver's read-then-clear pattern (main.rs:275-289) can
    be expressed end-to-end in P9.6.

  - `[[flags]]` families for INT_STATUS / INT_STATUS_ENABLE /
    INT_SIGNAL_ENABLE / SOFTWARE_RESET / PRESENT_STATE — bitflag
    newtypes with BitOr / BitAnd / Not replacing hand-rolled
    const-OR composition.

  - Multi-bit `fields` with `enum_values` for HOST_CONTROL.dma_sel
    ({Sdma, Adma2_32, Adma2_64}), POWER_CONTROL.bus_voltage
    ({V18, V30, V33}), CAPABILITIES.max_block_len
    ({B512, B1024, B2048}).

  - Per-field `access = "ro"` (Phase 8.2) on
    CLOCK_CONTROL.int_clk_stable — hardware-set status bit inside
    an otherwise-RW register; codegen suppresses the with_*
    setter so driver code that tries to write the stable bit
    fails to compile (used to silently no-op against hardware).

  - `verify_against = "lockjaw_types::sdhci"` + 28 [[verify_offsets]]
    bindings — every register's struct-field offset
    const_assert_eq's against the hand-written
    lockjaw_types::sdhci::SDHCI_* constant. Drift fails the
    build.

- `lockjaw-types/src/sdhci.rs` (mechanical sweep): all 25 SDHCI
  offset constants flip from `usize` to `u64`. The
  `offset_of!(Sdhci, name) as u64` cast in the generated
  verify_offsets const_asserts requires a u64 RHS; half-conversion
  is the worst state because the next agent adding a binding has
  to decide per-constant. All flipped together; rationale lives
  here in the commit message rather than per-constant in the
  source. Hand-written ADMA2_ATTR_* / SDHCI_CMD_* / SW_RST_* /
  SDHCI_TRNS_* / SDHCI_INT_* bit constants stay typed as u8/u16
  (not register offsets — out of scope for this sweep).

- `user/emmc2-driver/src/main.rs`: sdhci_read*/write* and
  poll_until_* helper signatures take `offset: u64` instead of
  `usize` — forced downstream of the constant flip. Helper bodies
  drop the redundant `as u64` casts; `(base + offset)` is now
  direct u64 arithmetic. Internal computation byte-identical;
  Pi-flash behavior unchanged. Helpers stay in place until
  P9.5/P9.6/P9.8 retire them per-width.

- `user/lockjaw-regs/src/sdhci.rs` (GENERATED, 1568 lines): the
  typed surface. 28/28 register offsets cross-checked. Highlights:
  - `set_transfer_mode_command(TransferMode, Command)` — swap-safe.
  - `argument2()` / `set_argument2()` / `modify_argument2()`
    delegate to the sysaddr cell.
  - `normal_int_status()` typed-snapshot read AND
    `clear_normal_int_status(NormalIntStatus)` typed mask write.
  - `ClockControl::int_clk_stable()` reader present; no
    `with_int_clk_stable` setter.

- `user/lockjaw-regs/src/lib.rs`: `pub mod sdhci;`.

- `xtask/src/gen_regs.rs` (+107 lines): adds the new `Access::Rw1c`
  variant and its emit paths. The pre-existing W1c variant emits
  clear-only accessors (correct for PL011 ICR-style registers with
  undefined read semantics); Rw1c emits both a typed snapshot read
  AND a typed clear (correct for SDHCI INT_STATUS where read and
  W1C share one offset). Cell type reused (W1c<T> already exposes
  both read() and clear(mask)); the difference is purely in the
  emitted accessor surface — Rw1c has no set_*/modify_* because
  value-store semantics don't fit a register whose writes clear
  bits. Aliased + test emit paths gain matching Rw1c arms.

  This was a substrate addition motivated by SDHCI INT_STATUS;
  no prior consumer existed. The pre-existing W1c comment
  incorrectly claimed "the readable counterpart [of SDHCI W1C] is
  a separate RO status register" — corrected to distinguish W1C
  (clear-only) from Rw1C (read+W1C).

Review rounds
-------------

Two substrate fixes landed as separate commits before this one,
both surfaced by P9.3's emit:

1. P9.1a (18245ff) — lifted the wildcard panic in
   `emit_device()`'s register-kind match. P9.0 and P9.1 added
   validation and accessor emit for combined_trigger and aliased
   but never wired explicit arms into the gating switch; SDHCI
   was the first consumer to trip the panic. Removing the wildcard
   also upgraded the gate from runtime panic to compile-time
   exhaustiveness — a future RegisterKind variant must add an arm
   here or the build fails.

2. P9.0b (0a7c7bd) — typed per-part newtypes for combined_trigger
   setters. Other-Claude flagged that the original signature
   `set_transfer_mode_command(transfer_mode: u16, command: u16)`
   admits silent positional swap (BCM2711 would execute the
   wrong-half-bits command without complaint). Same construction-
   safety class as the Adma2Descriptor (address, length) arg-order
   issue fixed in P9.2 review. Each part now gets a
   `#[repr(transparent)] pub struct PartName(pub uN)` newtype; the
   setter takes typed args; swap is a compile error.

Inline review fix in this commit:

3. Codex (medium) — the initial P9.3 staged diff modeled
   normal_int_status and error_int_status as `access = "w1c"`. The
   W1c codegen emits clear-only (no read accessor) on the wrong
   assumption that all W1C registers have a separate readable
   counterpart (true for PL011 RIS/MIS/ICR; false for SDHCI where
   read and W1C share one offset). The driver's existing
   read-then-clear pattern at main.rs:275-289 would have no typed
   surface to bind to in P9.6. Resolved by adding the
   `access = "rw1c"` access kind to gen_regs.rs and flipping the
   two registers to use it. This substrate addition lives in this
   commit (not split out as a separate substrate-fix commit)
   because there was no pre-existing P9.0/P9.1-style emit logic
   for Rw1c that P9.3 was tripping over — it's a new access mode
   that wouldn't have made sense to land before its first consumer.

Acceptance
----------

- `cargo xtask gen-regs --check`: clean (5 emitted, 0 parsed-only;
  cprman / fw_cfg / pl011 / virtio_mmio unchanged; sdhci.rs is
  the new file).
- `cargo test -p lockjaw-types --target aarch64-apple-darwin`:
  940 pass (no new tests in this commit; the sweep is type-only).
- `cargo test --manifest-path user/lockjaw-regs/Cargo.toml --target
  aarch64-apple-darwin`: 69 pass (+31 vs P9.2 — full SDHCI
  per-register layout / field-roundtrip / RO-field / aliased /
  combined_trigger / verify_against test fixture).
- `cargo build --manifest-path user/emmc2-driver/Cargo.toml
  --release`: clean (helper signature update is byte-identical
  internally).
- `make test`: 100/100 on GICv3 + GICv2.
- No compiler warnings.
- Pi flash: N/A (codegen-prep + driver helper signatures only;
  no behavioral driver changes).

+++ lockjaw-types/src/sdhci.rs
@@ -31,59 +31,59 @@ pub use crate::wire::sdhci::*;
 /// `SDMA_SYS_ADDR_REG` (0x000) — also serves as the 32-bit ADMA
 /// system address argument register. Read here only as part of
 /// dump-the-state diagnostics.
-pub const SDHCI_SYSADDR: usize = 0x000;
+pub const SDHCI_SYSADDR: u64 = 0x000;
 /// `BLOCK_SIZE_REG` (0x004) — bits[11:0] = block size in bytes.
-pub const SDHCI_BLOCK_SIZE: usize = 0x004;
+pub const SDHCI_BLOCK_SIZE: u64 = 0x004;
 /// `BLOCK_COUNT_REG` (0x006) — number of blocks to transfer.
-pub const SDHCI_BLOCK_COUNT: usize = 0x006;
+pub const SDHCI_BLOCK_COUNT: u64 = 0x006;
 /// `ARGUMENT_REG` (0x008) — command argument.
-pub const SDHCI_ARGUMENT: usize = 0x008;
+pub const SDHCI_ARGUMENT: u64 = 0x008;
 /// `TRANSFER_MODE_REG` (0x00c) + `COMMAND_REG` (0x00e) — combined
 /// 32-bit write triggers command issue.
-pub const SDHCI_TRANSFER_MODE: usize = 0x00c;
-pub const SDHCI_COMMAND: usize = 0x00e;
+pub const SDHCI_TRANSFER_MODE: u64 = 0x00c;
+pub const SDHCI_COMMAND: u64 = 0x00e;
 /// `RESPONSE_REG` (0x010..0x01f) — 4 × 32-bit response slots.
-pub const SDHCI_RESPONSE_0: usize = 0x010;
-pub const SDHCI_RESPONSE_1: usize = 0x014;
-pub const SDHCI_RESPONSE_2: usize = 0x018;
-pub const SDHCI_RESPONSE_3: usize = 0x01c;
+pub const SDHCI_RESPONSE_0: u64 = 0x010;
+pub const SDHCI_RESPONSE_1: u64 = 0x014;
+pub const SDHCI_RESPONSE_2: u64 = 0x018;
+pub const SDHCI_RESPONSE_3: u64 = 0x01c;
 /// `BUFFER_DATA_PORT_REG` (0x020) — PIO data window.
-pub const SDHCI_BUFFER_DATA_PORT: usize = 0x020;
+pub const SDHCI_BUFFER_DATA_PORT: u64 = 0x020;
 /// `PRESENT_STATE_REG` (0x024) — controller status snapshot.
-pub const SDHCI_PRESENT_STATE: usize = 0x024;
+pub const SDHCI_PRESENT_STATE: u64 = 0x024;
 /// `HOST_CONTROL_1` (0x028).
-pub const SDHCI_HOST_CONTROL: usize = 0x028;
+pub const SDHCI_HOST_CONTROL: u64 = 0x028;
 /// `POWER_CONTROL_REG` (0x029).
-pub const SDHCI_POWER_CONTROL: usize = 0x029;
+pub const SDHCI_POWER_CONTROL: u64 = 0x029;
 /// `BLOCK_GAP_CONTROL_REG` (0x02a).
-pub const SDHCI_BLOCK_GAP_CONTROL: usize = 0x02a;
+pub const SDHCI_BLOCK_GAP_CONTROL: u64 = 0x02a;
 /// `WAKEUP_CONTROL_REG` (0x02b).
-pub const SDHCI_WAKEUP_CONTROL: usize = 0x02b;
+pub const SDHCI_WAKEUP_CONTROL: u64 = 0x02b;
 /// `CLOCK_CONTROL_REG` (0x02c).
-pub const SDHCI_CLOCK_CONTROL: usize = 0x02c;
+pub const SDHCI_CLOCK_CONTROL: u64 = 0x02c;
 /// `TIMEOUT_CONTROL_REG` (0x02e).
-pub const SDHCI_TIMEOUT_CONTROL: usize = 0x02e;
+pub const SDHCI_TIMEOUT_CONTROL: u64 = 0x02e;
 /// `SOFTWARE_RESET_REG` (0x02f) — write `SW_RST_*` bits.
-pub const SDHCI_SOFTWARE_RESET: usize = 0x02f;
+pub const SDHCI_SOFTWARE_RESET: u64 = 0x02f;
 /// `NORMAL_INT_STATUS` (0x030).
-pub const SDHCI_NORMAL_INT_STATUS: usize = 0x030;
+pub const SDHCI_NORMAL_INT_STATUS: u64 = 0x030;
 /// `ERROR_INT_STATUS` (0x032).
-pub const SDHCI_ERROR_INT_STATUS: usize = 0x032;
+pub const SDHCI_ERROR_INT_STATUS: u64 = 0x032;
 /// `NORMAL_INT_STATUS_ENABLE` (0x034).
-pub const SDHCI_NORMAL_INT_STATUS_ENABLE: usize = 0x034;
+pub const SDHCI_NORMAL_INT_STATUS_ENABLE: u64 = 0x034;
 /// `ERROR_INT_STATUS_ENABLE` (0x036).
-pub const SDHCI_ERROR_INT_STATUS_ENABLE: usize = 0x036;
+pub const SDHCI_ERROR_INT_STATUS_ENABLE: u64 = 0x036;
 /// `NORMAL_INT_SIGNAL_ENABLE` (0x038).
-pub const SDHCI_NORMAL_INT_SIGNAL_ENABLE: usize = 0x038;
+pub const SDHCI_NORMAL_INT_SIGNAL_ENABLE: u64 = 0x038;
 /// `ERROR_INT_SIGNAL_ENABLE` (0x03a).
-pub const SDHCI_ERROR_INT_SIGNAL_ENABLE: usize = 0x03a;
+pub const SDHCI_ERROR_INT_SIGNAL_ENABLE: u64 = 0x03a;
 /// `CAPABILITIES` (0x040) — fixed device capabilities, low half.
-pub const SDHCI_CAPABILITIES: usize = 0x040;
+pub const SDHCI_CAPABILITIES: u64 = 0x040;
 /// `CAPABILITIES_HI` (0x044) — capabilities high half.
-pub const SDHCI_CAPABILITIES_HI: usize = 0x044;
+pub const SDHCI_CAPABILITIES_HI: u64 = 0x044;
 /// `SLOT_INT_STATUS_VERSION` (0x0fc) — slot status (low 16) +
 /// host controller version (high 16).
-pub const SDHCI_HOST_VERSION: usize = 0x0fe;
+pub const SDHCI_HOST_VERSION: u64 = 0x0fe;
 
 // ---------------------------------------------------------------------------
 // HOST_CONTROL_1 bits (offset 0x028)
@@ -187,7 +187,7 @@ pub const SDHCI_TRNS_MULTI: u16 = 0x0020;
 /// `ARGUMENT2_REG` (0x000) — argument register for Auto-CMD23. Aliases the
 /// `SDMA_SYS_ADDR_REG` location; the register's role is determined by which
 /// transfer mode is active (PIO + Auto-CMD23 here, SDMA in M6+).
-pub const SDHCI_ARGUMENT2: usize = 0x000;
+pub const SDHCI_ARGUMENT2: u64 = 0x000;
 
 // ---------------------------------------------------------------------------
 // HOST_CONTROL_1 DMA_SEL (bits[4:3], inside the existing SDHCI_HOST_CONTROL
@@ -215,7 +215,7 @@ pub const SDHCI_HOST_CTRL_DMA_SEL_ADMA2_32: u8 = 0b0001_0000;
 /// holds the high 32 bits (unused in this Lockjaw revision). DIFFERENT
 /// register from SDHCI_SYSADDR (0x000), which is the SDMA source/dest
 /// address (also aliased as ARGUMENT2 for Auto-CMD23 in M5).
-pub const SDHCI_ADMA_ADDRESS: usize = 0x058;
+pub const SDHCI_ADMA_ADDRESS: u64 = 0x058;
 
 // ---------------------------------------------------------------------------
 // ADMA2 descriptor — 32-bit address mode, per SDHCI v3 §1.13
